requirements.txt
openai==1.12.0
python-dotenv==1.0.0
datasets==2.16.0
gradio==4.15.0
